Extract interface for service which is used also outside of own context
